+++ lib/ruby_speech/grxml/grammar.rb
@@ -123,7 +123,8 @@ module RubySpeech
         traverse do |element|
           next unless element.is_a? Nokogiri::XML::Text
 
-          next if self.class.import(element.parent).is_a? Token
+          element_type = self.class.import(element.parent).class
+          next if [Token, Tag].include?(element_type)
 
           tokens = split_tokens(element).map do |string|
             Token.new.tap { |token| token << string }